replace guint by gsize for the glib2 version of irssi_ssl_read like it's defined in GIOFuncs for glib2 moved src/fe-text/utf8.* to src/fe-common/core changed get_utf8_char so it returns a status code and the unichar argument pointer to the value that it returned before if there were no errors, so you can check for a negative value an handle the error
git-svn-id: http://svn.irssi.org/repos/irssi/trunk@4091 dbcabf3a-b0e7-0310-adc4-f8d773084564

+++ src/core/network-openssl.c
@@ -277,7 +277,7 @@ static GIOStatus irssi_ssl_cert_step(GIOSSLChannel *chan)
 	return G_IO_STATUS_ERROR;
 }
 
-static GIOStatus irssi_ssl_read(GIOChannel *handle, gchar *buf, guint len, guint *ret, GError **gerr)
+static GIOStatus irssi_ssl_read(GIOChannel *handle, gchar *buf, gsize len, gsize *ret, GError **gerr)
 {
 	GIOSSLChannel *chan = (GIOSSLChannel *)handle;
 	gint err;
